refactor: renomenado o nome o modal
