+++ front/src/renderer/src/views/pages/Menu/components/Ingredients/DeleteIngredientModal/useDeleteIngredientModal.ts
